Add interpolated missing data
Missing data is now added in additional columns, 1 per probe. The
original, combined data is still in the first columns, so only
chart those if you want the previous behavior (data gaps).

I graph the additional lines in the same color, but with dotted
lines, to show the gaps.

The data is generated by looking for the next available data point
for the same probe, then averaging the difference. While this should
result in a straight line, it doesn't always, because of rounding
errors.